pyproject.toml
[project]
name = "biotopg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "accelerate>=1.6.0,<2",
    "chromadb>=1.0",
    "colorama>=0.4.6",
    "datasets<4",
    "dspy==2.6.27",
    "langchain-core<1.0",
    "langchain-text-splitters==0.0.1",
    "litellm<=1.67.4.post1",
    "networkx>=3.5",
    "nltk==3.8.1",
    "numpy>=2.2.3,<3",
    "pre-commit>=4.1.0,<5",
    "pydantic>=2.10.6,<3",
    "pydantic-core==2.33.1",
    "python-dotenv>=1.0.1,<2",
    "reuse>=5.0.2",
    "sentence-transformers>=3.4.1,<4",
    "tqdm>=4.0,<5",
]

[tool.setuptools]
packages = ["biotopg"]
